requirements.txt
beautifulsoup4==4.11.2
numpy==1.24.2
pandas==1.5.3
python_dateutil==2.8.2
requests==2.28.2
tqdm==4.65.0
